requirements.txt
discord.py>=1.7.1

python-dotenv>=0.15.0

praw>=7.2.0

mariadb>=1.0.6

toml>=0.10.2

aiomysql>=0.0.21

colorama>=0.4.4 